source: outline `source_file` generating/describing code from `model`
Now `source.model` mostly contains file logical content descriptors (`Type`
class hierarchy and auxiliary code).
